Fix code smells
