+++ molgenis-api-metadata/src/test/java/org/molgenis/api/metadata/v3/job/MetadataUpsertJobExecutionTest.java
@@ -53,6 +53,7 @@ public class MetadataUpsertJobExecutionTest extends AbstractSystemEntityTest {
     return attrs;
   }
 
+  @SuppressWarnings("squid:S2699") // Tests should include assertions
   @Test
   public void testSystemEntity() {
     internalTestAttributes(
